Update README.md
Human Resource is a very important for any business which drives the business to the path of success or bankruptcy if managed efficiently. In this project we will analyze the data regarding the recruitment, retention or retrenchment to improve the business outcomes. This analytic project will help the management to take data-driven decisions to promote the positive employees experience.